% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{DATA}
\alias{DATA}
\title{US GDP per capita and Inflation data.}
\format{A data frame with 287 rows and 3 variables:
\describe{
  \item{DATA}{Dates in quarters from 1948-01-01 until 2019-07-01}
  \item{GDP_PERCAPITA}{US GDP per capita, Chained 2012 Dollars, Seasonally Adjusted Annual Rate}
  \item{INFLATION}{Consumer Price Index for All Urban Consumers: All Items in U.S. City Average, Percent Change from Year Ago, Seasonally Adjusted}
}}
\source{
\url{https://fred.stlouisfed.org//}
}
\usage{
DATA
}
\description{
This dataset contains 287 quarterly observation of US GDP per capita and inflation percentage change from the year before, starting from 1948-01-01 and ending 2019-07-01.
}
\keyword{datasets}
