% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{ggdoubleplt}
\alias{ggdoubleplt}
\title{ggplot2 raw data visualization}
\usage{
ggdoubleplt()
}
\value{
This function gives as output a single window with the plot of the two series of the df, also adding them a tendency line.
}
\description{
This is a default function that is useful to visualize a better graph of the raw data using the package ggplot2.
}
\examples{
seriesplot <- ggdoubleplot()
}
