% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsDEF.R
\name{ggdoublesumplts}
\alias{ggdoublesumplts}
\title{Double summary plots}
\usage{
ggdoublesumplts(
  r = DATA,
  s = DATA$GDP_PERCAPITA,
  t1 = DATA$DATE,
  main1 = "Ts1",
  u = DATA,
  v = DATA$INFLATION,
  t2 = DATA$DATE,
  main2 = "Ts2"
)
}
\arguments{
\item{r}{The dataframe used.}

\item{s}{The series you want to visualize in x-axe}

\item{t1}{The dates you want to use to index your series.}

\item{main1}{The title you want to give to your plot.}

\item{u}{The other dataframe used.}

\item{v}{The other series you want to visualize in x-axe}

\item{t2}{The other dates you want to use to index your series.}

\item{main2}{The other title you want to give to your plot.}
}
\value{
This function returns the summary plots - raw data, ACF, PCF - of two time series.
}
\description{
This function works like 'ggsumplts' but permits to visualize two time series at the same time.
}
