% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsDEF.R
\name{ggsumplts}
\alias{ggsumplts}
\title{Summary plot}
\usage{
ggsumplts(
  r = DATA,
  s = DATA$GDP_PERCAPITA,
  t = DATA$DATE,
  main = "GDP per capita"
)
}
\arguments{
\item{r}{The dataframe used.}

\item{s}{The series you want to visualize in x-axe}

\item{t}{The dates you want to use to index your series.}

\item{main}{The title you want to give to your plot.}
}
\value{
This gunction returns the plot of the raw data, the ACF and the PACF of the desired data. By default it returns the GDP per capita of the default dataset of the package.
}
\description{
With this function you can visualize the raw data, the ACF and the PACF. By default the function uses the GDP data of the package.
}
\examples{
## You can try this    pltsINFL <- ggsumplts(r=DATA, s=DATA$INFLATION, t=DATA$DATE, main='INFLATION')

}
