% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{ggsumpltsINFL}
\alias{ggsumpltsINFL}
\title{ggplot2 summary visualization - INFL}
\usage{
ggsumpltsINFL()
}
\value{
This function returns the raw data, the acf and the pacf.
}
\description{
This is a default function used to visualize the raw data, acf and pacf of the GDP data.
}
