% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{I0_seriesGDP}
\alias{I0_seriesGDP}
\title{Differentiate time series - GDP}
\usage{
I0_seriesGDP(x = DATA$GDP_PERCAPITA, t = DATA$DATE)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
This function calculates the order of integration of your time series and gives you as output the differentiated one as a zoo object. By default this function uses the GDP data from 'DATA.csv'.
}
\description{
With this function is possible to differentiate a series to obtain the I(0) version. By default uses the GDP data from the default dataframe of the package 'DATA.CSV'.
}
\examples{
I0GDP <- I0_seriesGDP(x=GDP, t=time)
}
