% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{I0_seriesINFL}
\alias{I0_seriesINFL}
\title{Differentiate time series - INFLATION}
\usage{
I0_seriesINFL(x = DATA$INFLATION, t = DATA$DATE)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
This function calculates the order of integration of your time series and gives you as output the differentiated one as a zoo object. By default this function uses the Inflation data from 'DATA.csv'.
}
\description{
With this function is possible to differentiate a series to obtain the I(0) version. By default uses the Inflation data from the default dataframe of the package 'DATA.CSV'.
}
\examples{
I0INFL <- I0_seriesINFL(x=infl, t=time)
}
