% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{intorder}
\alias{intorder}
\title{Integration order}
\usage{
intorder(x)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}
}
\value{
This function returns the integration order of the series.
}
\description{
This function is used to calculate the order of integration of the series.
}
\examples{
IO_GDP <- intorderGDP(x=GDP)
}
