% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{intorderGDP}
\alias{intorderGDP}
\title{Integration order - GDP}
\usage{
intorderGDP(x = DATA$GDP_PERCAPITA)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}
}
\value{
This function returns the integration order of the series. By default this function uses the GDP data from 'DATA.csv'.
}
\description{
This function is used to calculate the order of integration of the series. By default uses the GDP data from the default dataframe of the package 'DATA.CSV'.
}
\examples{
IO_GDP <- intorderGDP(x=GDP)
}
