% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{intorderINFL}
\alias{intorderINFL}
\title{Integration order - Inflation}
\usage{
intorderINFL(x = DATA$INFLATION)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}
}
\value{
This function returns the integration order of the series. By default this function uses the Inflation data from 'DATA.csv'.
}
\description{
This function is used to calculate the order of integration of the series. By default uses the Inflation data from the default dataframe of the package 'DATA.CSV'.
}
\examples{
IO_INFL <- intorderINFL(x=inlflation)
}
