% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsDEF.R
\name{MAINPLOTS}
\alias{MAINPLOTS}
\title{Return the main plots of the package.}
\usage{
MAINPLOTS()
}
\value{
The function returns the main plots of TS1.
}
\description{
This function can be used to print the dataset and visualize the main plots of the package.
}
