% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{plot_I0GDP}
\alias{plot_I0GDP}
\title{Plot I(0) series - GDP}
\usage{
plot_I0GDP(x = DATA$GDP_PERCAPITA, t = DATA$DATE)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
This function gives as output the plot of the differentiated I(0) series. By default this function uses the GDP data from 'DATA.csv'.
}
\description{
With this function is possible to visualize the plot of a differentiate I(0) series. By default uses the GDP data from the default dataframe of the package 'DATA.CSV'.
}
\examples{
pI0GDP <- plot_I0GDP(x=GDP, t=time)
}
