% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{randar11}
\alias{randar11}
\title{Random ARIMA(1,1,0)}
\usage{
randar11(
  x = c(1, 1, 0),
  t = sample(100:1000, 1),
  phi = runif(1, min = -1, max = 1)
)
}
\arguments{
\item{x}{The vector which indicates the kind of ARMA model.}

\item{t}{The sample of the AR.}

\item{phi}{The coefficient of the AR.}
}
\value{
By default this function returns a random sample between 100 and 1000 observations of an ARMA(1,1,0) with a random coefficient between -1 and 1.
}
\description{
With this function - by default - it is possible to create a random AR(1) with a random set between 100 and 1000 obs, and a random coefficient.
}
\examples{
AR1 <- randar11(x=c(1,1,0), t=500, phi=-0.75)
}
