% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{randarma111}
\alias{randarma111}
\title{Random ARIMA(1,1,1)}
\usage{
randarma111(
  x = c(1, 1, 1),
  t = sample(100:1000, 1),
  phi = runif(1, min = -1, max = 1),
  teta = runif(1, min = -1, max = 1)
)
}
\arguments{
\item{x}{The vector which indicates the kind of ARMA model.}

\item{t}{The sample of the ARMA.}

\item{phi}{The coefficient of the AR.}

\item{teta}{The coefficient of the MA.}
}
\value{
By default this function returns a random sample between 100 and 1000 observations of an ARMA(1,1,1) with random coefficients between -1 and 1.
}
\description{
With this function - by default - it is possible to create a random ARMA(1,1) with a random set between 100 and 1000 obs, and random coefficients.
}
\examples{
ARMA11 <- randar11(x=c(0,1,1), t=500, phi=0.5, teta=-0.75)
}
