% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{randma11}
\alias{randma11}
\title{Random ARIMA(0,1,1)}
\usage{
randma11(
  x = c(0, 1, 1),
  t = sample(100:1000, 1),
  teta = runif(1, min = -1, max = 1)
)
}
\arguments{
\item{x}{The vector which indicates the kind of ARMA model.}

\item{t}{The sample of the MA.}

\item{teta}{The coefficient of the MA.}
}
\value{
By default this function returns a random sample between 100 and 1000 observations of an ARMA(0,1,1) with a random coefficient between -1 and 1.
}
\description{
With this function - by default - it is possible to create a random MA(1) with a random set between 100 and 1000 obs, and a random coefficient.
}
\examples{
MA1 <- randar11(x=c(0,1,1), t=500, teta=-0.75)
}
