% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sumplts}
\alias{sumplts}
\title{Time series' summary plots}
\usage{
sumplts(x, t)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
The function returns a series of three plots: The plot of the raw data, the ACF plot and the PACF plot.
}
\description{
This function is useful to visualize a summary - raw data, ACF and PACF - of your data in one window. WARNING: to visualize it click the icon above the 'refresh topic' one.
}
\examples{
sumPLOTS <- sumpltsGDP(x=data, t=time)
}
