% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sumpltsINFL}
\alias{sumpltsINFL}
\title{Inflation's summary plots}
\usage{
sumpltsINFL(x = DATA$INFLATION, t = DATA$DATE)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
The function returns a series of three plots: The plot of the raw data, the ACF plot and the PACF plot. As default it returns the plot of US inflation data.
}
\description{
This function is useful to visualize a summary - raw data, ACF and PACF - of your data in one window. WARNING: to visualize it click the icon above the 'refresh topic' one.
}
\examples{
sumPLOTS <- sumpltsINFL(x=data, t=time)
}
