% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{doublesumplts}
\alias{doublesumplts}
\title{Double column summary plots}
\usage{
doublesumplts(x = DATA$GDP_PERCAPITA, z = DATA$INFLATION, t = DATA$DATE)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{z}{The numeric vector of the other series you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
The function works creating two columns of graphs by which you can compare two series.
}
\description{
This function is useful to visualize a summary - raw data, ACF and PACF - of your two series in one window. WARNING: to visualize it click the icon above the 'refresh topic' one.
}
\examples{
twoseriesplot <- doublesumplts(x=GDP, z=INFL, t=time)
}
