% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsDEF.R
\name{ggdoubleplt}
\alias{ggdoubleplt}
\title{Plot raw data + tendency line}
\usage{
ggdoubleplt()
}
\value{
This function returns the plots of the two time series with a tendency line.
}
\description{
This default function is used to visualize the two time series of the 'DATA' dataframe
}
