% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{I0_series}
\alias{I0_series}
\title{Differentiate time series}
\usage{
I0_series(x, t)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
This function calculates the order of integration of your time series and gives you as output the differentiated one as a zoo object.
}
\description{
With this function is possible to differentiate a series to obtain the I(0) version.
}
\examples{
##do not run this    I0GDP <- I0_seriesGDP(x=GDP, t=time)
}
