% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsDEF.R
\name{intorder}
\alias{intorder}
\title{Integration order}
\usage{
intorder(x = DATA$GDP_PERCAPITA)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}
}
\value{
This function returns the integration order of the series.
}
\description{
This function is used to calculate the order of integration of the series. By default this function uses the US GDP per capita of the default dataset.
}
\examples{
##You can try this    IO_INFL <- intorder(x=DATA$INFLATION)
}
