% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{loaddata}
\alias{loaddata}
\title{Load the dataset}
\usage{
loaddata()
}
\value{
The function assigns the Data Frame of the file 'DATA.csv'
}
\description{
The function is used to download and assign a Data Frame from the file 'DATA.csv' which contains two series of observations downloaded on FRED. In particular they are the quarterly observations of US seasonally adjusted GDP per capita and Inflation, observed from 1947 to 2019.
}
