% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{MAINPLOTS}
\alias{MAINPLOTS}
\title{Return the main plots}
\usage{
MAINPLOTS()
}
\value{
The function returns the main plots of TS1.
}
\description{
This function can be used to upload the dataset and visualize the main plots of the package.
}
