% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsNEW.R
\name{plot_I0}
\alias{plot_I0}
\title{Plot I(0) series}
\usage{
plot_I0(x, t)
}
\arguments{
\item{x}{The numeric vector of the observations you want to visualize.}

\item{t}{The vector of the dates used to index the data.}
}
\value{
This function gives as output the plot of the differentiated I(0) series.
}
\description{
With this function is possible to visualize the plot of a differentiate I(0) series.
}
\examples{
pI0GDP <- plot_I0GDP(x=GDP, t=time)
}
